% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSpectra.R
\docType{data}
\name{BrainSpectra}
\alias{BrainSpectra}
\title{Brain Data from MetabolAnalyze}
\format{An object of class \code{list} of length 2.}
\usage{
BrainSpectra
}
\description{
A dataset containing the metabolomic of brains
}
\keyword{datasets}
