% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPPCA.R
\name{DPPCA}
\alias{DPPCA}
\title{Dynamic Probabilistic Principal Components model (DPPCA)}
\usage{
DPPCA(q, chain_output, prior_params, burn_in, thin, data_time)
}
\arguments{
\item{q}{Number of components.}

\item{chain_output}{Desired size of ouput chain.}

\item{prior_params}{Named list containing prior parameter specification. \cr The parameters to be supplied are: \cr
\eqn{\alpha}, \eqn{\beta}: prior of v2, a IG(\eqn{\alpha}/2, \eqn{\beta}/2) distribution. \cr
\eqn{\mu_\nu},  \eqn{\sigma^2_{\nu}}: prior of \eqn{\nu}, a N(\eqn{\mu_\nu},\eqn{\sigma^2_{\nu}}) distribution. \cr
\eqn{\mu_\phi},  \eqn{\sigma^2_\phi}: prior of \eqn{\phi}, a N_{[-1,1]}(\eqn{\mu_\phi},\eqn{\sigma^2_\phi}) distribution. \cr
\eqn{\alpha_V}, \eqn{\beta_V}: prior of V, a IG(\eqn{(\alpha_V)}/2, \eqn{(\beta_V)}/2) distribution. \cr
\eqn{\sigma^2_\mu}: prior of \eqn{\mu}, a N(0,\eqn{\sigma^2_\mu}) distribution. \cr
\eqn{\mu_\Phi},  \eqn{\sigma^2_{\Phi}}: prior of \eqn{\Phi}, a N_{[-1,1]}(\eqn{\mu_\Phi},\eqn{\sigma^2_\Phi}) distribution. \cr
\eqn{\Omega_m}: prior covariance matrix of \eqn{W_m}, a MVN(0, \eqn{\Omega_m}) distribution. \cr}

\item{burn_in}{length of burn-in period.}

\item{thin}{thinning to be performed on chain.}

\item{data_time}{List of M matrixes or data frames containing the observed data for each time point. \cr
The data frame should contain observations in rows and spectral bins in columns.}
}
\description{
DPCCA is a model for analysing longitudinal metabolomic data.
}
