% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dppca_results.R
\docType{data}
\name{dppca_results}
\alias{dppca_results}
\title{Dynamic Urine Data DPPCA output}
\format{An object of class \code{list} of length 12.}
\usage{
dppca_results
}
\description{
Dynamic Urine Data DPPCA output
}
\keyword{datasets}
