% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPPCA_top_loadings.R
\name{DPPCA_top_loadings}
\alias{DPPCA_top_loadings}
\title{Top n influential loading by time point}
\usage{
DPPCA_top_loadings(n, PC, W, cred_level = 0.95)
}
\arguments{
\item{n}{number of top n loadings to be returned}

\item{PC}{number of the principal component to consider}

\item{W}{chain of rotated loadings}

\item{cred_level}{cred_level percentage quantile-based credible interval}
}
\description{
Top n influential loading by time point
}
