% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LMMs.R
\name{fit_LMMs}
\alias{fit_LMMs}
\title{Fit Bayesian Linear Mixed Models to your top metabolites data}
\usage{
fit_LMMs(top, alpha, data)
}
\arguments{
\item{top}{object of class top_loadings}

\item{alpha}{(1-alpha) percentage credible interval for parameter significant assessment}

\item{data}{the original data set in list format}
}
\description{
Fit Bayesian Linear Mixed Models to your top metabolites data
}
