% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu_g.R
\name{get_mu_g}
\alias{get_mu_g}
\title{Obtain initial values of mean vector in MPPCA.}
\usage{
get_mu_g(data, g)
}
\arguments{
\item{data}{data to perform MPPCA on}

\item{g}{integer. Number of the group to get initial values for}
}
\value{
List
}
\description{
This function is used to get initial values for the mean vector \eqn{\mu} of a particular group in MPPCA.
}
\examples{


}
