% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_w_g.R
\name{get_w_g}
\alias{get_w_g}
\title{Obtain initial value of loading matrix in MPPCA.}
\usage{
get_w_g(data, g, q)
}
\arguments{
\item{data}{data to perform MPPCA on}

\item{g}{integer. Number of the group to get initial values for}

\item{q}{number of components}
}
\description{
Function used to obtain initial loadings matrix a group \code{g}.
Perfoms PPCA using only the observations in group \code{g}.
}
\seealso{
[PPCA()]
}
