% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_fit.R
\docType{data}
\name{lmm_fit}
\alias{lmm_fit}
\title{Dynamic Urine Data DPPCA output for LMM}
\format{An object of class \code{time_effects} of length 7.}
\usage{
lmm_fit
}
\description{
Dynamic Urine Data DPPCA output for LMM
}
\keyword{datasets}
