% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPCA.R
\name{MPPCA}
\alias{MPPCA}
\title{MPPCA}
\usage{
MPPCA(data, q_min, q_max, g_min, g_max, eps = 0.1, B = 100)
}
\arguments{
\item{data}{Data to perform MPPCA on}

\item{q_min}{Minumum number of components}

\item{q_max}{Maximum number of components}

\item{g_min}{Minimum number of clusters}

\item{g_max}{Maximum number of clusters}

\item{eps}{Tolerance for convergence of EM algorithm}

\item{B}{Number of bootstrap replicas for calculation of loadings sd}
}
\value{
List of outputs.
}
\description{
Fits Mixtures of Probabilistic Principal Components model to data.
}
