% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DPPCA_persistance.R
\name{plot.DPPCA_persistance}
\alias{plot.DPPCA_persistance}
\title{Plotting DPPCA persistance parameters}
\usage{
\method{plot}{DPPCA_persistance}(x, type = "histogram", ...)
}
\arguments{
\item{x}{object of class DPPCA_persistance}

\item{type}{either histogram or trace plot. Default is histogram.}

\item{...}{additional parameters}
}
\description{
Plotting DPPCA persistance parameters
}
