% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.influence.R
\name{plot.influence}
\alias{plot.influence}
\title{Plot influence of covariates}
\usage{
\method{plot}{influence}(x, ...)
}
\arguments{
\item{...}{}

\item{object}{of class influence_report}
}
\description{
Plot influence of covariates
}
