% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.influence.R
\name{plot.influence_report}
\alias{plot.influence_report}
\title{Plot influence of covariates}
\usage{
\method{plot}{influence_report}(x, PC, ...)
}
\arguments{
\item{x}{object of class influence_report}

\item{PC}{specific principal component}

\item{...}{additional arguments}
}
\description{
Plot influence of covariates
}
