% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loadings_chain.R
\name{plot.loadings_chain}
\alias{plot.loadings_chain}
\title{Trace plot of sampled loadings}
\usage{
\method{plot}{loadings_chain}(x, ...)
}
\arguments{
\item{x}{object of class loadings_chain}

\item{...}{additional arguments}
}
\description{
Chain of loadings for a randomly sampled spectral bin on a randomly sampled time point are plotted for convergence assessment.
}
