% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPPCA_BIC.R
\name{plot.MPPCA_BIC}
\alias{plot.MPPCA_BIC}
\title{Heatmap of BIC values for different g and q.}
\usage{
\method{plot}{MPPCA_BIC}(x, ...)
}
\arguments{
\item{x}{object of class MPPCA_BIC from MPPCA results}

\item{...}{a}
}
\description{
Heatmap of BIC values for different g and q.
}
