% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPPCA_loadings.R
\name{plot.MPPCA_loadings}
\alias{plot.MPPCA_loadings}
\title{Title}
\usage{
\method{plot}{MPPCA_loadings}(x, group, analysis = FALSE, PC = 1,
  n = 5, conf_level = 0.95, starting_n = 1, ...)
}
\arguments{
\item{x}{a}

\item{group}{a}

\item{analysis}{a}

\item{PC}{a}

\item{n}{a}

\item{conf_level}{a}

\item{starting_n}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
