% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPPCA_score.R
\name{plot.MPPCA_score}
\alias{plot.MPPCA_score}
\title{Title}
\usage{
\method{plot}{MPPCA_score}(x, x_axis_PC, y_axis_PC, clustering, labels,
  conf_level = 0.95, ...)
}
\arguments{
\item{x}{a}

\item{x_axis_PC}{a}

\item{y_axis_PC}{a}

\item{clustering}{a}

\item{labels}{a}

\item{conf_level}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
