% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PPCA.R
\name{plot.PPCA}
\alias{plot.PPCA}
\title{Title}
\usage{
\method{plot}{PPCA}(x, PC = 1, n = 5, conf_level = 0.95,
  starting_n = 1, ...)
}
\arguments{
\item{x}{a}

\item{PC}{a}

\item{n}{a}

\item{conf_level}{a}

\item{starting_n}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
