% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PPCA_diagnostic.R
\name{plot.PPCA_diagnostic}
\alias{plot.PPCA_diagnostic}
\title{Title}
\usage{
\method{plot}{PPCA_diagnostic}(x, max_ll = FALSE, ...)
}
\arguments{
\item{x}{A PPCA_diagnostic class (e.g. PPCA$diagnostic)}

\item{max_ll}{When \code{TRUE}, prints Likelihood Function Convergence Plot. When \code{FALSE}, print BIC and Proportion of Variance plot}

\item{...}{Other plots parameters}
}
\value{
Diagnostic plots
}
\description{
Title
}
\examples{
plot(PPCA(urine_data, q_max=2)$diagnostic)
}
