% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PPCA_score.R
\name{plot.PPCA_score}
\alias{plot.PPCA_score}
\title{Title}
\usage{
\method{plot}{PPCA_score}(x, x_axis_PC, y_axis_PC, labels,
  conf_level = 0.95, ...)
}
\arguments{
\item{x}{a}

\item{x_axis_PC}{a}

\item{y_axis_PC}{a}

\item{labels}{a}

\item{conf_level}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
