% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PPCA_significant.R
\name{plot.PPCA_significant}
\alias{plot.PPCA_significant}
\title{Title}
\usage{
\method{plot}{PPCA_significant}(x, ori_data, PC = 1, sample_no = 1,
  xlim, ...)
}
\arguments{
\item{x}{a}

\item{ori_data}{a}

\item{PC}{a}

\item{sample_no}{a}

\item{xlim}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
