% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scores_chain.R
\name{plot.scores_chain}
\alias{plot.scores_chain}
\title{Trace plot of sampled latent scores}
\usage{
\method{plot}{scores_chain}(x, ...)
}
\arguments{
\item{x}{object of class scores_chain}

\item{...}{additional arguments}
}
\description{
Chain of latent scores of a randomly sampled observation on a randomly sampled time point are plotted for convergence assessment.
}
