% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.time_effects.R
\name{plot.time_effects}
\alias{plot.time_effects}
\title{Plot top metabolites time effect}
\usage{
\method{plot}{time_effects}(x, ...)
}
\arguments{
\item{x}{object of class time_effects}

\item{...}{additional arguments}
}
\description{
Plot top metabolites time effect
}
