% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCA.R
\name{PPCA}
\alias{PPCA}
\title{Title}
\usage{
PPCA(data, covariates_data, q_min = 1, q_max = 10, eps = 0.01,
  max_it = 1000, B, conf_level = 0.95)
}
\arguments{
\item{data}{Metabolomic data}

\item{covariates_data}{Covariates from the sample attributes (e.g: weights, gender, regions, etc)}

\item{q_min}{Minimum number of principal component}

\item{q_max}{Maximum number of principal component}

\item{eps}{Epsilon criterion (decision threshold) for convergence}

\item{max_it}{Maximum iteration runs}

\item{B}{Number of Bootstrap runs}

\item{conf_level}{Confidence Level}
}
\value{
PPCA objects
}
\description{
Title
}
\examples{
PPCA(urine_data, q_max=2)
}
