% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCA_export_significant.R
\name{PPCA_export_significant}
\alias{PPCA_export_significant}
\title{Title}
\usage{
PPCA_export_significant(PPCA_significant,
  filename = "PPCA_Significant_x", shiny_path = NULL)
}
\arguments{
\item{PPCA_significant}{a}

\item{filename}{a}

\item{shiny_path}{a}
}
\value{
a
}
\description{
Title
}
