% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCA_one_q.R
\name{PPCA_one_q}
\alias{PPCA_one_q}
\title{Performs PPCA(PPCCA) in the data using a single value of \code{q}.}
\usage{
PPCA_one_q(data, covariates_data, q, B = 100, eps = 0.01,
  max_it = 1000, est.sd = FALSE, w.sd = FALSE)
}
\arguments{
\item{data}{data to perform MPPCA on}

\item{covariates_data}{covariate data if using PPCCA}

\item{q}{number of components}

\item{B}{number of bootstrap replicas to estimate standard devition of loadings}

\item{eps}{tolerance for EM algorithm convergence}

\item{max_it}{maximum number of EM algorithm iterations}

\item{est.sd}{Logical. If TRUE, standard deviation of loadings and covariate effects will be estimated via bootstrap. Default is FALSE}

\item{w.sd}{Logical. If TRUE, standard deviation of loadings will be estimated via bootstrap. Default is FALSE}

\item{w.sd}{Logical.}
}
\value{
list with PPCA (PPCCA) outputs
}
\description{
#' @importFrom stats prcomp qnorm sd glm gaussian
}
