% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling_type.R
\name{scale_data}
\alias{scale_data}
\title{Scales the data}
\usage{
scale_data(x, scale_type = "centering")
}
\arguments{
\item{x}{data}

\item{scale_type}{Type of scaling.}
}
\description{
Scales the data
}
