% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DPPCA_persistance.R
\name{summary.DPPCA_persistance}
\alias{summary.DPPCA_persistance}
\title{Summary of DPPCA persistance parameters}
\usage{
\method{summary}{DPPCA_persistance}(object, cred_level = 0.95, ...)
}
\arguments{
\item{object}{object of class DPPCA_persistance}

\item{cred_level}{cred_level percentage quantile-based credible interval}

\item{...}{additional parameters for summary}
}
\description{
Summary of DPPCA persistance parameters
}
