% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PPCA_significant.R
\name{summary.PPCA_significant}
\alias{summary.PPCA_significant}
\title{Summary of Significant Spectral bins}
\usage{
\method{summary}{PPCA_significant}(object, ...)
}
\arguments{
\item{object}{a}

\item{...}{a}
}
\value{
a
}
\description{
Summary of Significant Spectral bins
}
