% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.time_effects.R
\name{summary.time_effects}
\alias{summary.time_effects}
\title{Summary of top metabolites time effects}
\usage{
\method{summary}{time_effects}(object, ...)
}
\arguments{
\item{object}{object of class time_effects}

\item{...}{additional arguments}
}
\description{
Summary of top metabolites time effects
}
