% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{trace}
\alias{trace}
\title{Function to calculate the trace of a matrix.}
\usage{
trace(matrix)
}
\arguments{
\item{matrix}{}
}
\value{
Numeric. Trace value.
}
\description{
Function to calculate the trace of a matrix.
}
