% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urine_data.R
\docType{data}
\name{urine_data}
\alias{urine_data}
\title{Urine Data}
\format{An object of class \code{matrix} with 18 rows and 189 columns.}
\usage{
urine_data
}
\description{
A dataset containing the metabolomic of urines
}
\keyword{datasets}
