% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrineDynamic.R
\docType{data}
\name{UrineDynamic}
\alias{UrineDynamic}
\title{Dynamic Urine Data}
\format{An object of class \code{data.frame} with 241 rows and 193 columns.}
\usage{
UrineDynamic
}
\description{
A dataset containing the spectral bins at different time points of the urine metabolomics.
}
\keyword{datasets}
