% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrineSpectra.R
\docType{data}
\name{UrineSpectra}
\alias{UrineSpectra}
\title{Urine Data from MetabolAnalyze}
\format{An object of class \code{list} of length 2.}
\usage{
UrineSpectra
}
\description{
A dataset containing the metabolomic of urines
}
\keyword{datasets}
