% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DPPCA_persistance.R
\name{plot.DPPCA_persistance}
\alias{plot.DPPCA_persistance}
\title{Plotting DPPCA persistance parameters}
\usage{
\method{plot}{DPPCA_persistance}(x, ...)
}
\arguments{
\item{x}{object of class DPPCA_persistance}

\item{...}{additional parameters}
}
\description{
Plotting DPPCA persistance parameters
}
