% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPPCA_BIC.R
\name{plot.MPPCA_BIC}
\alias{plot.MPPCA_BIC}
\title{Heatmap of BIC values for different g and q.}
\usage{
\method{plot}{MPPCA_BIC}(bic_results)
}
\arguments{
\item{bic_results}{pbject of class MPPCA_BIC from MPPCA results}
}
\description{
Heatmap of BIC values for different g and q.
}
