% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.top_loadings.R
\name{plot.top_loadings}
\alias{plot.top_loadings}
\title{Plot top influential loadings by time point}
\usage{
\method{plot}{top_loadings}(x, ...)
}
\arguments{
\item{x}{object of class top_loadings}

\item{...}{additional arguments}
}
\description{
Plot top influential loadings by time point
}
